h-c-push: S3: Not removing remote files but update
Removing thousands of entries on S3/RadosGW is very slow, this
patch avoids removing or re-uploading entries when the files have
not changed. Files MD5 are compared with the local version.

At the end of the push, the old files/directories that where not
present in the local repository are removed. The runtime is now
heavily dependant on the number of files to re-upload or remove.
Pushing a new environment will be long but following update should
be very fast.