Replace external fact with YAML ENC config file
Initially, the scripts hpc-config-{push,apply} pushed/downloaded an
external fact YAML file on all nodes to propagate the cluster_name.

The hpc-config-{push,apply} scripts now propagate the new YAML
configuration file for the ENC cluster-node-classifier script instead
of the previous external fact YAML file.

If later hpc-config adopt the puppetmaster model, this file will stay
on puppetmaster side and it will not be required to propagate it on all
nodes anymore.