Add new script cluster-node-classifier for ENC
This commit introduces a new script cluster-node-classifier designed to
be used as ENC. Until now, Puppet-HPC relied on an external fact YAML
file propagated on all nodes to get the cluster_name. This approach is
not fully compatible with puppetmaster since facts are evaluted too
late on the master side for some processing.

A better approach is to use ENC for these parameters with new script
cluster-node-classifier. This script requires a YAML configuration file
as input. The previous external fact YAML file can be directly used
(w/o change) as the configuration file for this script.

The script also handles new optional prefix and areas which is a new
optional feature to split the configuration environment into multiples
parts.