h-c-{push,apply}: areas feature
This commit introduces a new optional areas feature to hpc-config. When
multiple areas are declared for a cluster (only one default area by
default), hpc-config-push splits the configuration environments into
multiples archives and subdirectories with sentivive data encrypted with
a encryption keys specific to each area. Then hpc-config-apply access
the configuration environment files of a specific area.

This feature aims to avoid an area from accessing the sensitive data
(ex: passwords) of another area. This feature is designed to harden the
security on HPC clusters but it is optional though.