Fix distro import
The silly try/except dance is difficult to read and even more so to
maintain. Instead we show the explicit dependency we have on distro
after python >= 3.7.
Also, distro.name on rhel returns 'Red Hat Enterprise Linux' which is
definitely not something we want: platform.dist()[0] returns
'redhat', hence the switch to distro.release_info().