push: checksum private files
For all files in the private directory tree, we create a checksum file
for each. This checksum will be served by Apache as X-Checksum-Sha256:
puppet will use it to see if a local file needs to be updated.

Without a checksum, the only comparison mechanism is the date and this
can lead to at least two issues:
1. your local file is not modified but you pushed a new puppet
   environment: because the file on the webserver is newer than the
   local one, puppet will tell you the timestamp of the remote one is
   newer. So it will overwrite it. Issue is you will have the same
   behavior everytime you push an environment.
2. your local file is modified and more recent than the one on the
   webserver: puppet will not update it. It will only update it if you
   push a new environment (so the file gets a "fresher" timestamp).